Implement mph algorithm extracted from MphIndexMatcher as MphMatcherGroup
